useAutocompleter in separate file
