+++ shared/teams/common/channels-widget.tsx
@@ -5,6 +5,7 @@ import * as Container from '../../util/container'
 import * as Constants from '../../constants/teams'
 import * as Types from '../../constants/types/teams'
 import ChannelPopup from '../team/settings-tab/channel-popup'
+import useAutocompleter from './use-autocompleter'
 
 type Props = {
   channels: Array<string>
@@ -66,78 +67,6 @@ const ChannelInputDesktop = ({onAdd, teamID}: ChannelInputProps) => {
   )
 }
 
-const useAutocompleter = (items: Array<string>, onSelect: (value: string) => void, filter: string) => {
-  const [selected, setSelected] = React.useState(0)
-  const filterLCase = filter.trim().toLowerCase()
-  const prevFilterLCase = Container.usePrevious(filterLCase)
-  React.useEffect(() => {
-    if (prevFilterLCase !== filterLCase) {
-      setSelected(0)
-    }
-  }, [setSelected, prevFilterLCase, filterLCase])
-  let itemsFiltered = filter ? items.filter(item => item.toLowerCase().includes(filterLCase)) : items
-  itemsFiltered = itemsFiltered.slice(0, 5)
-
-  const {popup, popupAnchor, setShowingPopup, showingPopup} = Kb.usePopup(
-    getAttachmentRef => (
-      <Kb.Overlay
-        attachTo={getAttachmentRef}
-        visible={showingPopup}
-        onHidden={() => setShowingPopup(false)}
-        matchDimension={true}
-        position="top center"
-        positionFallbacks={['bottom center']}
-      >
-        {itemsFiltered.map((item, idx) => (
-          <Kb.ClickableBox key={item} onMouseDown={() => onSelect(item)} onMouseOver={() => setSelected(idx)}>
-            <Kb.Box2
-              direction="horizontal"
-              fullWidth={true}
-              style={Styles.collapseStyles([styles.channelOption, selected === idx && styles.optionSelected])}
-            >
-              <Kb.Text type="BodySemibold" lineClamp={1}>
-                {item}
-              </Kb.Text>
-            </Kb.Box2>
-          </Kb.ClickableBox>
-        ))}
-      </Kb.Overlay>
-    ),
-    filterLCase + selected
-  )
-
-  const numItems = itemsFiltered.length
-  const selectedItem = itemsFiltered[selected]
-  const onKeyDown = React.useCallback(
-    evt => {
-      let diff = 0
-      switch (evt.key) {
-        case 'ArrowDown':
-          diff = 1
-          break
-        case 'ArrowUp':
-          diff = -1
-          break
-        case 'Enter':
-          onSelect(selectedItem)
-          break
-      }
-      let newSelected = selected + diff
-      if (newSelected >= numItems) {
-        newSelected = 0
-      } else if (newSelected < 0) {
-        newSelected = numItems - 1
-      }
-      if (newSelected !== selected) {
-        setSelected(newSelected)
-      }
-    },
-    [selected, setSelected, numItems, onSelect, selectedItem]
-  )
-
-  return {onKeyDown, popup, popupAnchor, setShowingPopup}
-}
-
 const ChannelInputMobile = ({onAdd, selected, teamID}: ChannelInputProps) => {
   const [showingPopup, setShowingPopup] = React.useState(false)
   const onComplete = (channelnames: Array<string>) => {
@@ -187,14 +116,10 @@ const styles = Styles.styleSheetCreate(() => ({
     paddingTop: Styles.globalMargins.xtiny,
   },
   channelDummyInputText: {color: Styles.globalColors.black_50},
-  channelOption: {...Styles.padding(4, 10, 2), backgroundColor: Styles.globalColors.white},
   container: {
     ...Styles.padding(Styles.globalMargins.tiny),
     backgroundColor: Styles.globalColors.blueGrey,
     borderRadius: Styles.borderRadius,
-  },
-  optionSelected: {
-    backgroundColor: Styles.globalColors.blueLighter2, // TODO transparent in dark mode?
   },
   pill: Styles.platformStyles({
     common: {
